Fixed a bug causing everything to be an orangish tint. Possibly because of floating point number

+++ Software-Rendering/src/Bitmap.h
@@ -10,7 +10,7 @@ class Bitmap
 		void Clear(Uint32 shade);
 		void SetPixel(int x,int y,Uint32 a,Uint32 r,Uint32 g,Uint32 b);
 		void CopyTexelToPixel(int texelXCoord,int texelYCoord,int pixelXCoord,int pixelYCoord,Bitmap* texture);
-		float GetPixel(int location){return m_pixels[location]; } 
+		int GetPixel(int location){return m_pixels[location]; } 
 		int GetWidth() { return m_width;}
 		int GetHeight() { return m_height;}
 		Uint32* GetPixels(){ return m_pixels; }
